+++ Tests/AdvancedOperationTests/Core/OperationStateTests.swift
@@ -27,23 +27,23 @@ import XCTest
 final class OperationStateTests: XCTestCase {
   func testTransition() {
     // ready
-    XCTAssertTrue(AsyncOperation.State.pending.canTransition(to: .executing))
-    XCTAssertTrue(AsyncOperation.State.pending.canTransition(to: .finished))
-    XCTAssertFalse(AsyncOperation.State.pending.canTransition(to: .pending))
+    XCTAssertTrue(AsyncOperation.State.ready.canTransition(to: .executing))
+    XCTAssertTrue(AsyncOperation.State.ready.canTransition(to: .finished))
+    XCTAssertFalse(AsyncOperation.State.ready.canTransition(to: .ready))
 
     // Executing
     XCTAssertFalse(AsyncOperation.State.executing.canTransition(to: .executing))
     XCTAssertTrue(AsyncOperation.State.executing.canTransition(to: .finished))
-    XCTAssertFalse(AsyncOperation.State.executing.canTransition(to: .pending))
+    XCTAssertFalse(AsyncOperation.State.executing.canTransition(to: .ready))
 
     // finished
     XCTAssertFalse(AsyncOperation.State.finished.canTransition(to: .executing))
     XCTAssertFalse(AsyncOperation.State.finished.canTransition(to: .finished))
-    XCTAssertFalse(AsyncOperation.State.finished.canTransition(to: .pending))
+    XCTAssertFalse(AsyncOperation.State.finished.canTransition(to: .ready))
   }
 
   func testDebugDescription() {
-    XCTAssertEqual(AsyncOperation.State.pending.debugDescription.lowercased(), "pending")
+    XCTAssertEqual(AsyncOperation.State.ready.debugDescription.lowercased(), "ready")
     XCTAssertEqual(AsyncOperation.State.executing.debugDescription.lowercased(), "executing")
     XCTAssertEqual(AsyncOperation.State.finished.debugDescription.lowercased(), "finished")
   }